requirements.txt
python-dotenv
websocket-client
websocket-server
requests